+++ リクエストをパースするWEBサーバーを作ってみる/.eslintignore
@@ -1,0 +1,7 @@
+dist/
+**/coverage/
+**/node_modules/
+**/*.min.js
+*.config.js
+*.config.ts
+.*lintrc.js
