+++ リクエストをパースするWEBサーバーを作ってみる/.gitignore
@@ -1,3 +1,2 @@
-/node_modules
-/coverage
-/dist
+node_modules
+dist
