--- リクエストをパースするWEBサーバーを作ってみる/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  transform: {
-    "^.+\\.ts$": "ts-jest",
-  },
-};
