+++ リクエストをパースするWEBサーバーを作ってみる/jest.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from "@jest/types";
+
+const config: Config.InitialOptions = {
+  verbose: true,
+  transform: {
+    "^.+\\.ts$": "ts-jest",
+  },
+};
+
+export default config;
