+++ リクエストをパースするWEBサーバーを作ってみる/package.json
@@ -2,6 +2,8 @@
   "name": "learn-express",
   "version": "1.0.0",
   "main": "index.ts",
+  "repository": "git@github.com:gn-t-k/learn-express.git",
+  "author": "gn-t-k <gntk1995422@gmail.com>",
   "license": "MIT",
   "scripts": {
     "eslint": "eslint",
@@ -12,33 +14,35 @@
     "start-ts": "ts-node ./src/index.ts",
     "dev": "ts-node-dev ./src/index.ts",
     "lint": "eslint --ext '.js,.ts' ./src/**",
+    "fix": "npm run -s format && npm run -s lint-fix",
+    "format": "prettier --write --loglevel=warn 'src/**/*.{js,ts,gql,graphql,json}'",
     "lint-fix": "eslint --fix --ext '.js,.ts' ./src/**",
+    "lint-conflict": "eslint-config-prettier 'src/**/*.{js,ts}'",
     "test": "jest --coverage",
     "test-watch": "jest --coverage --watch",
     "build": "tsc",
     "postinstall": "typesync"
   },
-  "dependencies": {
-    "express": "^4.17.1"
-  },
   "devDependencies": {
     "@types/eslint": "^7.2.6",
-    "@types/eslint-plugin-prettier": "^3.1.0",
     "@types/express": "^4.17.9",
-    "@types/node": "^14.14.19",
     "@types/prettier": "^2.1.6",
+    "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.11.1",
     "@typescript-eslint/parser": "^4.11.1",
     "eslint": "^7.17.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.0",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
+    "supertest": "^6.0.1",
     "ts-jest": "^26.4.4",
     "ts-node": "^9.1.1",
     "ts-node-dev": "^1.1.1",
     "typescript": "^4.1.3"
+  },
+  "dependencies": {
+    "express": "^4.17.1"
   }
 }
