+++ リクエストをパースするWEBサーバーを作ってみる/tsconfig.eslint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src/**/*.js", "src/**/*.ts"],
+  "exclude": ["node_modules"]
+}
